requirements.txt
# MLOps Pipeline Dependencies

# Azure ML
azure-ai-ml==1.10.0
azure-identity==1.14.0
azure-storage-blob==12.13.0
azure-mgmt-authorization==3.0.0
azureml-core==1.50.0
azureml-mlflow==1.50.0

# Data Processing
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.0.2
joblib==1.1.1

# ML Monitoring and Logging
mlflow==2.9.1
evidently==0.4.2

# Configuration
pyyaml==6.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
